Reduce calendar file size
